# contracts/agent-router/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: agent-router API
  version: 1.0.0
  summary: >-
    Planning-time and dispatch-time control plane for agent execution: which
    harness and model profile a task should run under, and which physical GPU
    should serve it.
  description: |
    `agent-router` answers two questions and reports on a third.

    - `POST /v1/route` - at planning time, given task metadata, which harness,
      model profile and effort should this attempt use, and what does it cost?
      The answer is stamped immutably into story frontmatter.
    - `POST /v1/place` - at dispatch time, given a stamped profile, which
      physical placement should serve it right now?
    - `POST /v1/capacity/heartbeat` and `GET /v1/status` - what compute exists,
      what state is it in, and what is loaded on it?

    ## What this service is not

    It **never proxies inference and never sits in the request path.** A caller
    receives names - a model profile, a placement - and reaches the model
    through agentgateway, which is the only model ingress (invariant 5).
    Nothing here returns a URL to call.

    It **never mutates Flux-managed gateway configuration.** Gateway topology
    is Git-managed and stays that way; the router expresses placement per
    request through the `x-placement` header, not by rewriting CRs.

    It **introduces no CRD.** Catalog data lives in a Git-managed ConfigMap
    until a concrete requirement forces otherwise, and that requirement gets
...

...

...

...

...

...

...

...

...
servers:
...
security:
  - callerBearer: []
paths:
  /v1/route:
    post:
      operationId: route
      tags: [routing]
      summary: Decide harness, model profile and effort for a task.
      description: |
        Planning-time decision. Takes story/task metadata and returns an
        `ExecutionProfile` that is stamped IMMUTABLY into story frontmatter
        (invariant 2). A worker may not re-choose its model mid-attempt;
        re-routing is a new attempt with a new call here.

        **Metered is default-deny (invariant 4).** If the best option would
        spend billable budget and the request did not set `allow_metered: true`,
        the router does not return it. It returns the best non-metered option
        with a `metered_denied` note (HTTP 200), or - when no non-metered option
        exists - HTTP 409 `metered_denied` with no profile at all. It never
        returns the metered choice with a warning attached, and a client that
        reacts by resending with `allow_metered: true` has defeated the control:
        that authorization is a human decision, per attempt, and is recorded.

        **`allow_metered: true` is intent, not authority.** The flag asks for a
        billable candidate to be considered; whether one may be returned
        depends on the AUTHENTICATED CALLER independently holding permission to
        approve billable spend. The ordinary automation principal - the
        planning flow that stamps stories, and every machine caller like
        it - does not hold it, and a request from one carrying the flag is
        answered `403 metered_authorization_required`, not served. A control a
        client can grant itself is not a control.

        **`placement_required` on the response tells the client whether an
        inference placement is still owed.** True means the profile is local
        and the client must call `/v1/place` and send the returned
        `x-placement`; false means it is vendor-hosted, resolved by the
        harness, and there is no placement to make. Without the field a client
        has to keep its own copy of which profiles are local, or guess from the
        name prefix - both of which go stale silently when the catalog moves.

        It is a MODEL-INFERENCE placement, and only that. Nothing here decides
        where the agent process itself runs, or starts anything: running the
        session is the ADE's, per the ownership boundary above.

        **`local-agent` is never emitted** (invariant 16). It is a reserved,
        unsupported harness value; the catalog marks it `refuse-to-emit` and the
        response schema's harness enum cannot express it.

        Side-effect free and safe to retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
            examples:
              contained-code-change:
                summary: Low-risk single-module change
                description: Yields examples/execution-profile/local-code-standard.json.
                value:
                  story_id: STORY-034-19
                  title: Add a DNS record for the gateway hostname
                  summary: >-
                    Add the missing LAN DNS record so the gateway hostname
                    resolves without a --resolve override. One manifest, one
                    test.
                  touched_paths:
                    - kubernetes/apps/network/
                  ambiguity: low
                  blast_radius: low
                  context_size: 12000
                  volume_hint: low
                  tags: [iac]
                  repo: prox-ops
                  requester: bmad-pm
              security-tagged:
                summary: Security-tagged work, hard-excludes the unrestricted profile
                description: Yields examples/execution-profile/security-tagged-excludes-unrestricted.json.
                value:
                  story_id: STORY-035-21
                  title: Implement machine identity and tool-level authz
                  summary: >-
                    Autonomous workers authenticate as themselves to federated
                    MCP and see only their permitted tools.
                  touched_paths:
                    - kubernetes/apps/mcp/
                  ambiguity: high
                  blast_radius: critical
                  context_size: 90000
                  volume_hint: high
                  tags: [security, iam, secrets]
                  repo: prox-ops
                  requester: bmad-pm
              metered-not-authorized:
                summary: No allow_metered, so a billable option is withheld
                description: Yields examples/execution-profile/metered-denied-substituted.json.
                value:
                  story_id: STORY-035-15
                  title: Cross-hardware benchmark matrix
                  summary: Same metric set across every placement, one committed table.
                  ambiguity: medium
                  blast_radius: low
                  context_size: 40000
                  volume_hint: high
                  tags: [docs]
                  repo: prox-ops
                  requester: bmad-pm
                  allow_metered: false
      responses:
        '200':
          description: |
            A decision. Carries a `metered_denied` note when a billable option
            was withheld and this is the best remaining non-metered one.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionProfile'
              examples:
                local-code-standard:
                  summary: Free local profile, two fallbacks
                  externalValue: ./examples/execution-profile/local-code-standard.json
                security-tagged:
                  summary: Unrestricted profile hard-excluded by tags
                  externalValue: ./examples/execution-profile/security-tagged-excludes-unrestricted.json
                metered-denied:
                  summary: Billable option withheld, non-metered substitute returned
                  externalValue: ./examples/execution-profile/metered-denied-substituted.json
                docs-low-risk:
                  summary: Non-code local profile, no fallbacks on purpose
                  externalValue: ./examples/execution-profile/docs-low-risk-cluster-only.json
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          description: |
            The request set `allow_metered: true` but this caller is not
            authorized to approve billable spend. Nothing was returned and
            nothing was spent. The caller's obligation is to escalate to a
            principal that holds the authority - never to retry, and never to
            retry under a different credential.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                metered-authorization-required:
                  summary: Metered intent from a principal without metered authority
                  externalValue: ./examples/errors/metered-authorization-required.json
        '409':
          description: |
            The request cannot be served as framed. `metered_denied` when every
            option is billable and none was authorized; `no_eligible_profile`
            when nothing satisfies the constraints; `catalog_version_stale` when
            the caller pinned a catalog the router is no longer serving.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                metered-denied:
                  externalValue: ./examples/errors/metered-denied-no-alternative.json
                no-eligible-profile:
                  externalValue: ./examples/errors/no-eligible-profile.json
                catalog-version-stale:
                  externalValue: ./examples/errors/catalog-version-stale.json
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/CatalogUnavailable'
...
  /v1/capacity/heartbeat:
    post:
      operationId: heartbeat
      tags: [capacity]
      summary: Report edge worker capacity and state.
      description: |
        Edge to router. The body is the Edge Worker Contract's heartbeat
        payload (`edge/EDGE-WORKER-CONTRACT.md` section 1), unchanged.

        Authentication is a per-node bearer credential, and it is bound to the
        `node` field: a valid credential presenting a different node id is
        rejected 403 `node_identity_mismatch`, because a credential that could
        heartbeat as any node could withdraw a healthy placement or keep a dead
        one advertised.

        **TLS is required on this path, always.** An edge worker is off-cluster
        by definition, and the credential it presents is long-lived, so this
        call uses the `https://` server and the transport rules in the API
        description above. There is no plaintext variant of the edge heartbeat.

        **A heartbeat may narrow static authority; it may never expand it:**

            effective capability = Git/catalog authority INTERSECT heartbeat observation

        The payload is an observation, and observations subtract. A node can
        report itself busy, drained, cold or short of VRAM, and the router acts
        on that at once. It cannot report itself INTO a model, a capability, a
        context size or any other authority the Git-managed catalog does not
        already give that placement. `active_model` counts as warm only when
        that model is catalog-authorized for the placement; a claim with no
        catalog counterpart is a reconciliation alarm and is ignored for
        eligibility. This is not a model-authorization bypass either way -
        authorization is the per-model CEL policy, which never reads a
        heartbeat - but an unbounded self-report would let a compromised node
        bias placement toward itself, and placement integrity is worth
        defending on its own.

        **State ownership is local.** The host decides its own state; the
        cluster never forces a node into `SERVING`. The response acknowledges
        what the router recorded and tells the node how often to report - it is
        not a command.

        **A router-side failure never ends the heartbeat loop.** Recording a
        self-reported state needs no catalog, so this endpoint returns only the
        transient `catalog_unavailable` and never the permanent
        `catalog_schema_unsupported`: an `abort` obligation aimed at an edge
        node means "stop reporting", and a node that obeyed it would stay
        invisible after an operator had fixed the router. The loop terminates
        only on a failure the node itself has to fix - a rejected credential, a
        node identity that is not its own.

        This is withdrawal mechanism 1 of 3 (see README "Withdrawal
        semantics"): `state: INTERACTIVE` or `DRAINING` marks the placement
        ineligible immediately, on receipt, with no timeout involved.
      security:
        - edgeNodeBearer: []
      # TLS ONLY, mechanically. The top-level `servers` list carries a
      # plaintext ClusterIP entry for in-cluster callers, and an operation
      # inherits that list unless it overrides it. This endpoint is called
      # from OFF-CLUSTER by definition, with a long-lived per-node
      # credential, so it overrides: a generated client for this operation
      # can only be pointed at the TLS server, and the plaintext one is not
      # an option it is offered. Prose saying "use HTTPS here" is not the
      # same thing.
      servers:
        - url: https://{edge_host}
          description: >-
            The only server this operation may be called on. See "Transport"
            in the API description for the certificate requirements: real CA,
            hostname SAN, no insecureSkipVerify, dedicated key.
          variables:
            edge_host:
              default: agent-router.<lan-domain>
              description: >-
                DNS name the certificate is issued for, provisioned with it.
                The placeholder is not a resolvable name.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Heartbeat'
            examples:
              available:
                summary: Idle, model warm
                externalValue: ./examples/heartbeat/state-available.json
              serving:
                summary: Handling inference
                externalValue: ./examples/heartbeat/state-serving.json
              draining:
                summary: Finishing in-flight work, refusing new
                externalValue: ./examples/heartbeat/state-draining.json
              interactive:
                summary: A human has claimed the GPU
                externalValue: ./examples/heartbeat/state-interactive.json
              offline:
                summary: Explicit self-report before shutdown
                externalValue: ./examples/heartbeat/state-offline.json
              unmeasured-laptop:
                summary: Unmeasured hardware on battery
                externalValue: ./examples/heartbeat/unmeasured-laptop-on-battery.json
      responses:
        '202':
          description: |
            Recorded. `eligible_for_placement` is the router's answer to what
            the node just reported, so a node can see immediately that it has
            been taken out of the selection path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeartbeatAck'
              examples:
                accepted:
                  summary: Available node, still eligible
                  value:
                    accepted: true
                    node: <edge-node-1>
                    observed_state: AVAILABLE
                    eligible_for_placement: true
                    next_heartbeat_after_seconds: 30
                    offline_after_seconds: 90
                    catalog_version: sha256:2fd681e0e988bf6be94b8923b1485a0aa174a4e66aec580a2d383c764b60e229
                withdrawn:
                  summary: Interactive node, marked ineligible on receipt
                  value:
                    accepted: true
                    node: <edge-node-1>
                    observed_state: INTERACTIVE
                    eligible_for_placement: false
                    next_heartbeat_after_seconds: 30
                    offline_after_seconds: 90
                    catalog_version: sha256:2fd681e0e988bf6be94b8923b1485a0aa174a4e66aec580a2d383c764b60e229
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          description: Missing or invalid bearer credential.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthenticated:
                  externalValue: ./examples/errors/heartbeat-unauthenticated.json
        '403':
          description: >-
            The credential is valid but is not bound to the node named in the
            payload. The heartbeat is discarded; no capacity state changes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                node-identity-mismatch:
                  externalValue: ./examples/errors/heartbeat-node-identity-mismatch.json
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: >-
            The router cannot record capacity right now. Only the TRANSIENT
            `catalog_unavailable` is served here - never
            `catalog_schema_unsupported`, whose `abort` obligation would tell a
            node to stop reporting for good over a condition on the router's
            side of the wire. Back off and heartbeat again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                catalog-unavailable:
                  summary: Transient - retry on the next interval
                  externalValue: ./examples/errors/catalog-unavailable.json
  /v1/status:
    get:
      operationId: status
      tags: [introspection]
      summary: Catalog, compute providers, health and warm-model state.
      description: |
        The router's whole read surface: profiles to physical models to
        eligible placements, plus each placement's state, readiness and last
        heartbeat. JSON is the contract; an HTML rendering may be added later
        without a contract change.

        Each edge placement re-exposes its most recent heartbeat VERBATIM in
        the Edge Worker Contract's own shape, so there is exactly one heartbeat
        shape in this contract rather than a second one to drift from it.

        `runtime.endpoint` is visible here because this is a status surface and
        R9 makes that field observational metadata. Reading it is its purpose.
        Routing on it is forbidden - for the router, and for every consumer of
        this response.

        Answering this call never probes anything. A KServe placement reports
        `readiness: "unknown"` rather than being pinged, because a status
        endpoint that woke predictors to fill a field in would be exactly the
        mechanism invariant 7 forbids.
      responses:
        '200':
          description: Current catalog and capacity view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                today:
                  summary: >-
                    Catalog 1.3.0 as it stands - cachyos-7900xtx selectable
                    and heartbeating, local-code-standard with a second
                    physical candidate, and the four retired local models
                    gone from local-code-fast/local-general/local-unrestricted
                  externalValue: ./examples/status/status-today.json
                pre-edge:
                  summary: >-
                    Historical - catalog 1.1.0 before cachyos-7900xtx was
                    brought up, one usable placement
                  externalValue: ./examples/status/status-pre-edge-1.1.0.json
                restarted-unseen:
                  summary: >-
                    Today's catalog, seconds after a router restart. The edge
                    node is enrolled and may well be healthy, but nothing has
                    been observed yet, so it is `unseen` and not eligible. Read
                    with `capacity_state: learning`, which is what makes an
                    unavailable answer provisional rather than final.
                  externalValue: ./examples/status/status-restarted-unseen.json
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: >-
            No usable catalog is loaded - either none yet
            (`catalog_unavailable`, transient, retry) or one the router refused
            because its `schema_version` major is not implemented
            (`catalog_schema_unsupported`, permanent, abort). It refuses rather
            than best-effort parsing an unknown shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                catalog-unavailable:
                  summary: Transient - nothing loaded yet, retry
                  externalValue: ./examples/errors/catalog-unavailable.json
                catalog-schema-unsupported:
                  summary: Permanent - unimplemented catalog schema_version, abort
                  externalValue: ./examples/errors/catalog-schema-unsupported.json
components:
  securitySchemes:
    callerBearer:
      type: http
      scheme: bearer
      description: >-
        Caller credential for planning and dispatch clients (the BMAD planning
        flow, an operator, a conformance client). Distinct from the edge node
        credential: a client that can ask for a placement must not be able to
        report capacity.
    edgeNodeBearer:
      type: http
      scheme: bearer
      description: >-
        Per-node edge credential, bound to one `node` identifier. Provisioned
        from 1Password - `ExternalSecret` cluster-side, `op read` host-side -
        and never committed. Presenting it for a different node is 403, not a
        silent accept. It is long-lived and it travels off-cluster, so it is
        only ever sent over TLS: see "Transport" in the API description.
  schemas:
    ExecutionProfile:
      $ref: './schemas/execution-profile.schema.json'
...
    Error:
      $ref: './schemas/error.schema.json'
...
  responses:
    InvalidRequest:
      description: The request body is malformed or a field is out of range.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid-request:
              externalValue: ./examples/errors/invalid-request.json
    Unauthenticated:
      description: >-
        Missing or invalid CALLER bearer credential. The heartbeat endpoint has
        its own 401 because it authenticates with the separate per-node edge
        credential, and an example naming the wrong endpoint teaches the wrong
        error handler.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            unauthenticated:
              summary: Caller credential rejected; no endpoint named, since every caller endpoint serves this
              externalValue: ./examples/errors/caller-unauthenticated.json
    RateLimited:
      description: >-
        Too many requests. The router is a planning-time service; a caller
        hitting it per inference request should be caching placements for their
        `ttl_seconds` instead.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            rate-limited:
              externalValue: ./examples/errors/rate-limited.json
    InternalError:
      description: >-
        Unexpected router failure. Both decision endpoints are side-effect
        free, so a retry is safe; a persistent failure fails the attempt rather
        than degrading it onto a different profile.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            internal-error:
              externalValue: ./examples/errors/internal-error.json
...